Update example scripts using differential update

+++ example/clone_intact.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+sync_update=true
+
 if [ ! `which psql` ] ; then
 
  echo "psql: command not found..."
@@ -79,10 +81,20 @@ if [ ! -d $XML_DIR ] ; then
 
 fi
 
-psql -d $DB_NAME -U $DB_USER -f $DB_SCHEMA --quiet
+relations=`psql -d $DB_NAME -U $DB_USER -c "\d" | wc -l`
+
+if [ $sync_update != "true" ] || [ $relations = "0" ] ; then
+ psql -d $DB_NAME -U $DB_USER -f $DB_SCHEMA --quiet
+fi
 
 WORK_DIR=pg_work
-CSV_DIR=$WORK_DIR/csv
+
+if [ $sync_update != "true" ] ; then
+ CSV_DIR=$WORK_DIR/csv
+else
+ MD5_DIR=chk_sum_pgsql
+fi
+
 ERR_DIR=$WORK_DIR/err
 
 err_file=$ERR_DIR/all_err
@@ -90,17 +102,32 @@ err_file=$ERR_DIR/all_err
 rm -rf $WORK_DIR
 
 mkdir -p $WORK_DIR
-mkdir -p $CSV_DIR
+
+if [ $sync_update != "true" ] ; then
+ mkdir -p $CSV_DIR
+fi
+
 mkdir -p $ERR_DIR
 
 err_file=$ERR_DIR/all_err
 
-java -classpath ../xsd2pgschema.jar xml2pgcsv --xsd $XSD_SCHEMA --xml $XML_DIR --csv-dir $CSV_DIR --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+if [ $sync_update != "true" ] ; then
+
+ java -classpath ../xsd2pgschema.jar xml2pgcsv --xsd $XSD_SCHEMA --xml $XML_DIR --csv-dir $CSV_DIR --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+
+else
+
+ java -classpath ../xsd2pgschema.jar xml2pgsql --xsd $XSD_SCHEMA --xml $XML_DIR --sync $MD5_DIR --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+
+fi
 
 if [ $? = 0 ] && [ ! -s $err_file ] ; then
 
  rm -f $err_file
- rm -rf $CSV_DIR
+
+ if [ $sync_update != "true" ] ; then
+  rm -rf $CSV_DIR
+ fi
 
 else
